Fixed Most Code and Removed Filter Package
The filter package (meant to add a general filteable interface was too
convoluted to continue with. There were point where the code was pass a
concrete slice of objects into a function that utilized pointers and
type assertions- only to be returned as a slice of interface values. It
was really dumb. I have reinstated the old code methodology while
somewhat change up the directories to promote a cleaner project.